requirements.txt
pandas
typer
typing_extensions
google-cloud-bigquery
